backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
google-generativeai==0.5.4
motor==3.4.0
pydantic==2.7.1
python-multipart==0.0.9
websockets==12.0
httpx==0.27.0
